Fixed style issues as reported by IntelliJ
git-svn-id: https://svn.apache.org/repos/asf/commons/proper/compress/trunk@1650766 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/commons/compress/archivers/zip/ZipArchiveOutputStream.java
@@ -1142,7 +1142,7 @@ public class ZipArchiveOutputStream extends ArchiveOutputStream {
 
     private byte[] createCentralFileHeader(ZipArchiveEntry ze) throws IOException {
 
-        final long lfhOffset = offsets.get(ze).longValue();
+        final long lfhOffset = offsets.get(ze);
         final boolean needsZip64Extra = hasZip64Extra(ze)
                 || ze.getCompressedSize() >= ZIP64_MAGIC
                 || ze.getSize() >= ZIP64_MAGIC
